+++ e2e/journeys/manageCases/caseProgression/completeTheOrder/completeTheOrder.ts
@@ -1,15 +1,15 @@
 import { Browser, Page } from "@playwright/test";
-import { Helpers } from "../../../../../common/helpers.ts";
-import { AdminEditAndApproveAnOrder1Page } from "../../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder1Page.ts";
-import { AdminEditAndApproveAnOrder4Page } from "../../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder4Page.ts";
-import { AdminEditAndApproveAnOrder21Page } from "../../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder21Page.ts";
-import { AdminEditAndApproveAnOrder22Page } from "../../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder22Page.ts";
-import { AdminEditAndApproveAnOrder23Page } from "../../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder23Page.ts";
-import { AdminEditAndApproveAnOrderSubmitPage } from "../../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrderSubmitPage.ts";
-import config from "../../../../../config.ts";
-import { createOrderFL401Options } from "../../../../../common/types.ts";
-import { ApplicationJourneysCheckGatekeeperJudgeCOOrder } from "../application-journeys-check-gatekeeper-judgeCO-order.ts";
-import { jsonDatas } from "../../../../../common/solicitorCaseCreatorHelper.ts";
+import { createOrderFL401Options } from "../../../../common/types.ts";
+import { jsonDatas } from "../../../../common/solicitorCaseCreatorHelper.ts";
+import { ApplicationJourneysCheckGatekeeperJudgeCOOrder } from "../createACaseUpToServiceOfApplicationState/application-journeys-check-gatekeeper-judgeCO-order.ts";
+import { Helpers } from "../../../../common/helpers.ts";
+import config from "../../../../config.ts";
+import { AdminEditAndApproveAnOrder1Page } from "../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder1Page.ts";
+import { AdminEditAndApproveAnOrder4Page } from "../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder4Page.ts";
+import { AdminEditAndApproveAnOrder21Page } from "../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder21Page.ts";
+import { AdminEditAndApproveAnOrder22Page } from "../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder22Page.ts";
+import { AdminEditAndApproveAnOrder23Page } from "../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrder23Page.ts";
+import { AdminEditAndApproveAnOrderSubmitPage } from "../../../../pages/manageCases/caseProgression/completeTheOrder/adminEditAndApproveAnOrderSubmitPage.ts";
 
 interface CompleteTheOrderParams {
   page: Page;
